integers for iteration

+++ utils.py
@@ -45,7 +45,8 @@ def inline_estim(vec, fit_model='hk', bins='knuth', inv='spm', winsize=1000.,
     # Windows along-track
     #--------------------------------------------------------------------------
     x = np.arange(vec.size) #vector index
-    xa = x[:x.size-winsize:sampling] #windows starting coordinate
+    #xa = x[:x.size-winsize:sampling] #windows starting coordinate
+    xa = x[:np.int(x.size-winsize):np.int(sampling)] #windows starting coordinate
     xb = xa + winsize-1 #window end coordinate
     if xb[-1] > x[-1]: xb[-1] = x[-1] #cut last window in limb
     xo = [val+(xb[i]-val)/2. for i, val in enumerate(xa)]
